+++ prefect/flows/ingest_cdc_profiles_web_to_gcs.py
@@ -1,10 +1,8 @@
-from typing import Literal, List, Dict
+from datetime import datetime
+from typing import Dict, List, Literal
 
 import polars as pl
-
-import requests
-
-from datetime import datetime
+from utils.chess_dot_com_api import request_from_chess_dot_com_public_api
 
 
 def check_title_abbrv(
@@ -27,25 +25,6 @@ def check_title_abbrv(
     if title_abbrv not in allowed_title_abbrvs:
         error_message = f"Title abbreviation is not valid: {title_abbrv}"
         raise ValueError(error_message)
-
-
-def request_from_chess_dot_com_public_api(
-    api_endpoint_suffix: str, headers: Dict = {"User-Agent": "default@domain.com"}
-) -> Dict:
-    """Function which queries the public Chess.com API and returns the JSON response."""
-    # Construct the API request URL
-    api_endpoint_url = f"https://api.chess.com/pub/{api_endpoint_suffix}"
-
-    # Query API
-    response = requests.get(api_endpoint_url, headers=headers)
-
-    # Check for empty response body
-    response.raise_for_status()
-    if response.status_code != 204:
-        return response.json()
-    raise ValueError(
-        f"API response for URL {api_endpoint_url} is empty and has status code 204."
-    )
 
 
 def get_titled_players_usernames(
